+++ src/theme/utils/isLightTheme.ts
@@ -1,0 +1,5 @@
+import * as vscode from 'vscode'
+
+export const isLightTheme = (activeColorTheme: vscode.ColorTheme): boolean => {
+  return activeColorTheme.kind === 1 || activeColorTheme.kind === 4
+}
